requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.0
pip==25.3
pygame==2.6.1
PyOpenGL==3.1.10
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
